builder,publisher: support master_doc
Introduce the ability to configure a space's homepage based off a
documentation's `master_doc` configuration [1].

In default configurations, this extension publishes documents either
in the root of a configured space or in a parent page. This requires
both an automated document set (what this extension will publish)
and a manual document set (a user preparing a space's homepage or
parent page). A new option "confluence_master_homepage" has been
introduced that when set "True" and the Sphinx configuration defines
a `master_doc` document, the target space's homepage will be
adjusted to the published page that maps to the `master_doc` entry.

Note: This is a partial implementation; this does not work for
publishers using the REST API. See included TODO for more
information.

[1]: http://www.sphinx-doc.org/en/stable/config.html#confval-master_doc

Signed-off-by: James Knight <james.d.knight@live.com>